+++ packages/runtime/test/util/assemble-state.test.ts
@@ -6,9 +6,9 @@ import assembleState from '../../src/util/assemble-state';
 
 test('with no arguments, returns a basic state object', (t) => {
   const initial = undefined;
-  const data = undefined;
+  const defaultState = undefined;
   const config = undefined;
-  const result = assembleState(initial, config, data);
+  const result = assembleState(initial, config, defaultState);
   t.deepEqual(result, {
     configuration: {},
     data: {},
@@ -17,9 +17,9 @@ test('with no arguments, returns a basic state object', (t) => {
 
 test('strict: ignores initial state', (t) => {
   const initial = { x: 22 };
-  const data = undefined;
+  const defaultState = undefined;
   const config = undefined;
-  const result = assembleState(initial, config, data, true);
+  const result = assembleState(initial, config, defaultState, true);
   t.deepEqual(result, {
     configuration: {},
     data: {},
@@ -28,9 +28,9 @@ test('strict: ignores initial state', (t) => {
 
 test('strict: ignores initial state except references', (t) => {
   const initial = { references: [] };
-  const data = undefined;
+  const defaultState = undefined;
   const config = undefined;
-  const result = assembleState(initial, config, data, true);
+  const result = assembleState(initial, config, defaultState, true);
   t.deepEqual(result, {
     references: [],
     configuration: {},
@@ -40,9 +40,9 @@ test('strict: ignores initial state except references', (t) => {
 
 test('non-strict: includes initial state', (t) => {
   const initial = { x: 22 };
-  const data = undefined;
+  const defaultState = undefined;
   const config = undefined;
-  const result = assembleState(initial, config, data, false);
+  const result = assembleState(initial, config, defaultState, false);
   t.deepEqual(result, {
     x: 22,
     configuration: {},
@@ -52,10 +52,10 @@ test('non-strict: includes initial state', (t) => {
 
 test('merges default and initial data objects', (t) => {
   const initial = { data: { x: 1 } };
-  const data = { y: 1 };
+  const defaultState = { data: { y: 1 } };
   const config = undefined;
 
-  const strict = assembleState(initial, config, data, true);
+  const strict = assembleState(initial, config, defaultState, true);
   t.deepEqual(strict, {
     configuration: {},
     data: {
@@ -65,16 +65,16 @@ test('merges default and initial data objects', (t) => {
   });
 
   // Ensure the same behaviour in non-strict mode
-  const nonStrict = assembleState(initial, config, data, false);
+  const nonStrict = assembleState(initial, config, defaultState, false);
   t.deepEqual(strict, nonStrict);
 });
 
 test('Initial data is prioritised over default data', (t) => {
   const initial = { data: { x: 1 } };
-  const data = { x: 2 };
+  const defaultState = { data: { x: 2 } };
   const config = undefined;
 
-  const strict = assembleState(initial, config, data, true);
+  const strict = assembleState(initial, config, defaultState, true);
   t.deepEqual(strict, {
     configuration: {},
     data: {
@@ -82,16 +82,16 @@ test('Initial data is prioritised over default data', (t) => {
     },
   });
 
-  const nonStrict = assembleState(initial, config, data, false);
+  const nonStrict = assembleState(initial, config, defaultState, false);
   t.deepEqual(strict, nonStrict);
 });
 
 test('merges default and initial config objects', (t) => {
   const initial = { configuration: { x: 1 } };
-  const data = undefined;
+  const defaultState = undefined;
   const config = { y: 1 };
 
-  const strict = assembleState(initial, config, data, true);
+  const strict = assembleState(initial, config, defaultState, true);
   t.deepEqual(strict, {
     configuration: {
       x: 1,
@@ -101,16 +101,16 @@ test('merges default and initial config objects', (t) => {
   });
 
   // Ensure the same behaviour in non-strict mode
-  const nonStrict = assembleState(initial, config, data, false);
+  const nonStrict = assembleState(initial, config, defaultState, false);
   t.deepEqual(strict, nonStrict);
 });
 
 test('configuration overrides initialState.configuration', (t) => {
   const initial = { configuration: { x: 1 } };
-  const data = undefined;
+  const defaultState = undefined;
   const config = { x: 2 };
 
-  const strict = assembleState(initial, config, data, true);
+  const strict = assembleState(initial, config, defaultState, true);
   t.deepEqual(strict, {
     configuration: {
       x: 2,
@@ -119,6 +119,6 @@ test('configuration overrides initialState.configuration', (t) => {
   });
 
   // Ensure the same behaviour in non-strict mode
-  const nonStrict = assembleState(initial, config, data, false);
+  const nonStrict = assembleState(initial, config, defaultState, false);
   t.deepEqual(strict, nonStrict);
 });